requirements.txt
Flask>=2.2.2
Flask-Cors>=3.0.10
py-algorand-sdk>=2.0.0
pyteal>=0.23.0
beaker-pyteal>=0.5.3
